+++ app/helpers/wilday_ui/application_helper.rb
@@ -2,6 +2,5 @@ module WildayUi
   module ApplicationHelper
     include WildayUi::Components::ButtonHelper
     include WildayUi::JavascriptHelper
-    include WildayUi::ThemeHelper
   end
 end