+++ app/javascript/wilday_ui/controllers/index.js
@@ -3,6 +3,7 @@ import ButtonController from "./button_controller";
 import DropdownController from "./dropdown_controller";
 import ClipboardController from "./clipboard_controller";
 import ConfirmationController from "./confirmation_controller";
+import TooltipController from "./tooltip_controller";
 // Initialize Stimulus
 const application = Application.start();
 window.Stimulus = application;
@@ -12,6 +13,7 @@ application.register("button", ButtonController);
 application.register("dropdown", DropdownController);
 application.register("clipboard", ClipboardController);
 application.register("confirmation", ConfirmationController);
+application.register("tooltip", TooltipController);
 // Debug check to ensure Stimulus is loaded
 // if (window.Stimulus) {
 //   console.log("✅ Stimulus is loaded and initialized.");